# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: LiveChat API
  version: 1.0.0
  description: >-
    API REST e contrato WebSocket/STOMP do LiveChat. Datas no formato
    date-time são representadas em UTC.
...
paths:
  /users/register:
    post:
      tags: [Usuários]
      operationId: registerUser
      summary: Cadastra um usuário
      description: Cria uma conta e armazena a senha com BCrypt. Não inicia uma sessão.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '200':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/login:
    post:
      tags: [Usuários]
      operationId: loginUser
      summary: Autentica um usuário
      description: Retorna um JWT de acesso válido por duas horas e um refresh token válido por sete dias.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Credenciais válidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/refresh:
    post:
      tags: [Usuários]
      operationId: refreshUserTokens
      summary: Renova os tokens
      description: Consome o refresh token atual e devolve um novo par. O token consumido não pode ser reutilizado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens renovados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/password-reset/request:
    post:
      tags: [Usuários]
      operationId: requestPasswordReset
      summary: Solicita recuperação de senha
      description: >-
        Sempre responde 202 para não revelar se o e-mail está cadastrado. Quando a conta existe,
        envia um link de uso único para o e-mail do usuário; o token expira em 15 minutos por padrão.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '202':
          description: Solicitação recebida, independentemente da existência da conta.
  /users/password-reset/confirm:
    post:
      tags: [Usuários]
      operationId: confirmPasswordReset
      summary: Define uma nova senha com o token recebido por e-mail
      description: >-
        Consome o token uma única vez, altera a senha e invalida os JWTs e refresh tokens emitidos
        anteriormente para a conta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '204':
          description: Senha alterada e sessões anteriores invalidadas.
        '400':
          $ref: '#/components/responses/BadRequest'
  /users/profile-update/confirm:
    post:
      tags: [Usuários]
      operationId: confirmProfileUpdate
      summary: Confirma uma atualização pendente do perfil
      description: >-
        Consome o token de uso único enviado ao e-mail anterior e aplica o novo nome e/ou e-mail.
        Quando o e-mail muda, JWTs e refresh tokens anteriores são invalidados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateConfirmRequest'
      responses:
        '204':
          description: Perfil atualizado.
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users:
    get:
      tags: [Usuários]
      operationId: listUsers
      summary: Lista os usuários cadastrados
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuários encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    delete:
      tags: [Usuários]
      operationId: deleteUser
      summary: Desativa e anonimiza a própria conta
      description: >-
        O identificador deve pertencer ao usuário autenticado. A conta é mantida como tombstone para
        preservar o histórico, mas perde acesso, dados pessoais, vínculos sociais e memberships. Os
        servidores são transferidos ao membro ativo mais antigo; servidores sem sucessor são removidos.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Conta anonimizada, desativada e credenciais revogadas.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/search:
    get:
      tags: [Usuários]
      operationId: searchUsersByEmail
      summary: Busca um usuário por e-mail
      description: A implementação atual realiza correspondência exata de e-mail.
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          description: E-mail completo do usuário procurado.
          schema:
            type: string
            format: email
          example: ana@example.com
      responses:
        '200':
          description: Lista vazia ou usuário encontrado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/password-reset:
    post:
      tags: [Usuários]
      operationId: requestAuthenticatedPasswordReset
      summary: Solicita troca de senha para a conta autenticada
      description: >-
        Envia ao e-mail atual o mesmo link usado na recuperação de senha. O cliente não informa o
        endereço; ele é obtido do usuário autenticado. A nova senha é definida em
        `/users/password-reset/confirm`.
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Solicitação recebida.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      tags: [Usuários]
      operationId: getCurrentUser
      summary: Obtém o usuário autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados públicos da conta autenticada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Usuários]
      operationId: requestProfileUpdate
      summary: Solicita alteração do nome e/ou e-mail
      description: >-
        Não altera o perfil imediatamente. Cria uma solicitação de uso único e envia o link de
        confirmação ao e-mail atual, inclusive quando o endereço solicitado é diferente.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '202':
          description: Atualização pendente criada e confirmação enviada ao e-mail atual.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de acesso retornado por `/users/login` ou `/users/refresh`.
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: UUID da conta.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requisição inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: JWT ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Usuário autenticado sem permissão para a operação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso inexistente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: O recurso já existe ou conflita com o estado atual.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Erro interno da aplicação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    UserRegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Ana Silva
        email:
          type: string
          format: email
          example: ana@example.com
        password:
          type: string
          format: password
          example: senha-segura
    UserLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: ana@example.com
        password:
          type: string
          format: password
          example: senha-segura
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Token opaco retornado no login ou na última renovação.
          example: fQ9wVjW8P0Qm0xW6vYgR3nY9kF2uZcT8P4sL7aE1bN0
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: ana@example.com
    PasswordResetConfirmRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 43
          maxLength: 43
          description: Token opaco de uso único recebido no link enviado por e-mail.
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: nova-senha-segura
    ProfileUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Novo nome exibido, normalizado com remoção de espaços nas extremidades.
          example: Ana Silva
        email:
          type: string
          format: email
          maxLength: 254
          description: Novo e-mail, normalizado para minúsculas.
          example: ana.nova@example.com
    ProfileUpdateConfirmRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 43
          maxLength: 43
          description: Token opaco de uso único enviado ao e-mail anterior.
    AuthTokens:
      type: object
      required: [name, token, expiresIn, refreshToken, refreshExpiresIn]
      properties:
        name:
          type: string
          example: Ana Silva
        token:
          type: string
          description: JWT de acesso.
          example: eyJhbGciOiJIUzI1NiJ9...
        expiresIn:
          type: string
          format: date-time
        refreshToken:
          type: string
          description: Token usado exclusivamente em `/users/refresh`.
        refreshExpiresIn:
          type: string
          format: date-time
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Ana Silva
        email:
          type: string
          format: email
          example: ana@example.com
...
    Error:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: JWT ausente ou inválido